pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "saddler"
dynamic = ["version"]
description = "Agent harness assembler and lifecycle manager"
authors = [{ name = "Chen Wang", email = "hopefulwc@gmail.com" }]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
keywords = ["agent", "cli", "harness", "ai", "orchestration", "saddler"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
  "typer>=0.12.0",
  "filelock>=3.16.0",
  "psutil>=7.2.2",
  "pyyaml>=6.0.3",
  "pydantic>=2.13.3",
]

[project.scripts]
saddler = "saddler.cli:main"

[project.urls]
Homepage = "https://github.com/hopeful0/saddler"
Repository = "https://github.com/hopeful0/saddler"
Issues = "https://github.com/hopeful0/saddler/issues"

[dependency-groups]
dev = [
  "prek>=0.3.8",
  "pytest>=8.4.2",
  "ruff>=0.15.11",
]

[tool.hatch.version]
source = "vcs"
